+++ api/urls.py
@@ -1,31 +1,12 @@
-from . import views
+from .v1.views import invalid_endpoint
 
 try:
-    from django.urls import re_path as url
+    from django.urls import include, re_path as url
 except ImportError:
-    from django.conf.urls import url
-
-
-app_name = 'api'
-
-_series = '^series/(?P<slug>[^/]+)'
-_volume = '%s/(?P<vol>[^/]+)/' % _series
-_authors = '^authors'
-_artists = '^artists'
-_groups = '^groups'
+    from django.conf.urls import include, url
 
 urlpatterns = [
-    url(r'^$', views.invalid_endpoint, name='invalid'),
-    url('^releases/?$', views.all_releases, name='releases'),
-    url('^series/?$', views.all_series, name='series_root'),
-    url('%s/?$' % _series, views.series, name='series'),
-    url('%s/?$' % _volume, views.volume, name='volume'),
-    url('%s/(?P<num>[^/]+)/?$' % _volume, views.chapter, name='chapter'),
-    url('%s/?$' % _authors, views.all_people, name='author_root'),
-    url('%s/(?P<p_id>[^/]+)/?$' % _authors, views.person, name='author'),
-    url('%s/?$' % _artists, views.all_people, name='artist_root'),
-    url('%s/(?P<p_id>[^/]+)/?$' % _artists, views.person, name='artist'),
-    url('%s/?$' % _groups, views.all_groups, name='groups_root'),
-    url('%s/(?P<g_id>[^/]+)/?$' % _groups, views.group, name='group'),
+    url(r'^$', invalid_endpoint, name='root'),
+    url(r'^v1/', include('api.v1.urls')),
 ]