+++ reader/views.py
@@ -7,16 +7,19 @@ from .models import Chapter, Series
 
 def directory(request):
     return render(request, 'directory.html', {
-        'all_series': Series.objects.all().order_by('title')
+        'all_series': Series.objects.prefetch_related(
+            'chapters').order_by('title')
     })
 
 
 def series(request, slug=None):
     try:
-        _series = Series.objects.get(slug=slug)
+        _series = Series.objects.prefetch_related(
+            'chapters__groups', 'artists', 'categories',
+            'authors', 'aliases').get(slug=slug)
     except ObjectDoesNotExist:
         raise Http404
-    if _series.chapters.count() == 0 and not request.user.is_staff:
+    if not (_series.chapters.count() or request.user.is_staff):
         return render(request, 'error.html', {
             'error_message': 'Sorry. This series is not yet available.',
             'error_status': 403
@@ -32,7 +35,8 @@ def chapter_page(request, slug=None, vol=0, num=0, page=1):
     if page == 0:
         raise Http404
     chapters = {
-        'all': Chapter.objects.filter(series__slug=slug),
+        'all': Chapter.objects.prefetch_related(
+            'series__categories', 'pages').filter(series_id=slug),
         'curr': None,
         'prev': None,
         'next': None
@@ -44,7 +48,7 @@ def chapter_page(request, slug=None, vol=0, num=0, page=1):
     chapters['next'] = next_in_order(chapters['curr'])
     chapters['prev'] = prev_in_order(chapters['curr'])
     all_pages = chapters['curr'].pages.all()
-    if page > all_pages.count():
+    if page > len(all_pages):
         raise Http404
     return render(request, 'chapter.html', {
         'all_chapters': chapters['all'].reverse(),